# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Amestris API
  version: 1.0.0
  description: API para gestión de alquimistas, materiales, misiones y transmutaciones.

servers:
  - url: http://localhost:8080/api
    description: Dev

paths:

  /healthz:
    get:
      summary: Health check del servicio
      tags: [System]
      responses:
        "200":
          description: Estado básico del backend
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  db:     { type: string }
                  checkedAt: { type: string, format: date-time }

  /readyz:
    get:
      summary: Readiness check (DB lista)
      tags: [System]
      responses:
        "200":
          description: El backend está listo para recibir tráfico

  /metrics:
    get:
      summary: Métricas del servicio (formato tipo Prometheus)
      tags: [System]
      responses:
        "200":
          description: Métricas de rendimiento y jobs

  # ============ AUTH ============

  /auth/register:
    post:
      summary: Registrar usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:     { type: string, example: "Roy Mustang" }
                email:    { type: string, example: "roy@amestris.gov" }
                password: { type: string, example: "1234" }
                role:     { type: string, enum: [SUPERVISOR, ALCHEMIST], example: SUPERVISOR }
      responses:
        "201":
          description: Usuario creado y login automático
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'

  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:    { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'

  /auth/me:
    get:
      summary: Datos del usuario autenticado
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Usuario actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /auth/refresh:
    post:
      summary: Rotar tokens (refresh)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshInput'
      responses:
        "200":
          description: Tokens rotados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'

  /auth/logout:
    post:
      summary: Revocar refresh actual (logout)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jti]
              properties:
                jti: { type: string, example: "ab12cd34..." }
      responses:
        "200":
          description: Logout exitoso

  # ============ MATERIALS ============

  /materials:
    get:
      summary: Listar materiales
      tags: [Materials]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de materiales
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Material'
                  page:     { type: integer }
                  pageSize: { type: integer }
                  total:    { type: integer }
    post:
      summary: Crear material
      tags: [Materials]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialInput'
      responses:
        "201":
          description: Creado

  /materials/{id}:
    put:
      summary: Actualizar material
      tags: [Materials]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialInput'
      responses:
        "200":
          description: Actualizado
    delete:
      summary: Eliminar material
      tags: [Materials]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Eliminado

  # ============ MISSIONS ============

  /missions:
    get:
      summary: Listar misiones
      tags: [Missions]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de misiones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Mission' }
    post:
      summary: Crear misión
      tags: [Missions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionCreate'
      responses:
        "201":
          description: Creada

  /missions/{id}:
    get:
      summary: Obtener detalle de misión
      tags: [Missions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detalle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
    put:
      summary: Actualizar misión
      tags: [Missions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdate'
      responses:
        "200":
          description: Actualizada
    delete:
      summary: Eliminar misión
      tags: [Missions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Eliminada

  # ============ TRANSMUTATIONS ============

  /transmutations:
    get:
      summary: Listar transmutaciones
      tags: [Transmutations]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transmutation'
                  page:     { type: integer }
                  pageSize: { type: integer }
                  total:    { type: integer }
    post:
      summary: Crear transmutación (con control de stock)
      tags: [Transmutations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransmutationCreate'
      responses:
        "201":
          description: Creada

  /transmutations/{id}:
    delete:
      summary: Eliminar transmutación (restaura stock)
      tags: [Transmutations]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Sin contenido

  # ============ ALCHEMISTS & AUDITS ============

  /alchemists:
    get:
      summary: Listar alquimistas
      tags: [Alchemists]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Alchemist' }

  /alchemists/{id}:
    get:
      summary: Ver detalle de alquimista
      tags: [Alchemists]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detalle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alchemist'
    put:
      summary: Actualizar alquimista
      tags: [Alchemists]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlchemistUpdate'
      responses:
        "200":
          description: Actualizado
    delete:
      summary: Eliminar alquimista
      tags: [Alchemists]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Eliminado

  /audits:
    get:
      summary: Listar registros de auditoría
      tags: [Audits]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de auditorías
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Audit' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    User:
      type: object
      properties:
        id:    { type: integer }
        name:  { type: string }
        email: { type: string }
        role:  { type: string, enum: [SUPERVISOR, ALCHEMIST] }

    AuthTokens:
      type: object
      properties:
        token:   { type: string, description: "alias de access para compatibilidad" }
        user:
          $ref: '#/components/schemas/User'
        access:  { type: string }
        refresh: { type: string }
        jti:     { type: string }
        exp:     { type: integer, description: "exp del access (unix)" }

    RefreshInput:
      type: object
      required: [refresh, jti]
      properties:
        refresh: { type: string }
        jti:     { type: string }

    Material:
      type: object
      properties:
        id:       { type: integer }
        name:     { type: string }
        quantity: { type: number }
        unit:     { type: string }
        rarity:   { type: string, nullable: true }
        notes:    { type: string, nullable: true }

    MaterialInput:
      type: object
      properties:
        name:     { type: string }
        quantity: { type: number }
        unit:     { type: string }
        rarity:   { type: string, nullable: true }
        notes:    { type: string, nullable: true }

    Mission:
      type: object
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: string, nullable: true }
        status:      { type: string }
        assignedAlchemistId:   { type: integer, nullable: true }
        assignedAlchemistName: { type: string, nullable: true }

    MissionCreate:
      type: object
      required: [title]
      properties:
        title:       { type: string }
        description: { type: string, nullable: true }
        status:      { type: string, example: "OPEN" }
        assignedAlchemistId: { type: integer, nullable: true }

    MissionUpdate:
      type: object
      properties:
        title:       { type: string }
        description: { type: string, nullable: true }
        status:      { type: string }
        assignedAlchemistId: { type: integer, nullable: true }

    Transmutation:
      type: object
      properties:
        id:           { type: integer }
        title:        { type: string }
        materialId:   { type: integer }
        materialName: { type: string, nullable: true }
        missionId:    { type: integer, nullable: true }
        missionTitle: { type: string, nullable: true }
        quantityUsed: { type: number }
        result:       { type: string, nullable: true }
        createdAt:    { type: string, format: date-time }

    TransmutationCreate:
      type: object
      required: [title, materialId, quantityUsed]
      properties:
        title:        { type: string }
        materialId:   { type: integer }
        missionId:    { type: integer, nullable: true }
        quantityUsed: { type: number }
        result:       { type: string, nullable: true }

    Alchemist:
      type: object
      properties:
        id:   { type: integer }
        name: { type: string }
        rank: { type: string, nullable: true }

    AlchemistUpdate:
      type: object
      properties:
        name: { type: string }
        rank: { type: string }

    Audit:
      type: object
      properties:
        id:       { type: integer }
        action:   { type: string }
        entity:   { type: string }
        entityId: { type: integer }
        meta:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }

security:
  - bearerAuth: []